Change pilot->cpu to an int. This avoids the player being told they have -0 CPU and can't fly their ship.

+++ src/pilot.h
@@ -276,8 +276,8 @@ typedef struct Pilot_ {
    int tsy;          /**< current sprite y position, calculated on update. */
 
    /* Properties. */
-   double cpu;       /**< Amount of CPU the pilot has left. */
-   double cpu_max;   /**< Maximum amount of CPU the pilot has. */
+   int cpu;       /**< Amount of CPU the pilot has left. */
+   int cpu_max;   /**< Maximum amount of CPU the pilot has. */
    double crew;      /**< Crew amount the player has (display it as (int)floor(), but it's analogue. */
    double cap_cargo; /**< Pilot's cargo capacity. */
